Extract PMD configuration
